he arreglado algun id, cuando no se especifica que tipo de integer es, por defecto es biginteger y yo declaraba que era integer, asi que ahora lo cambie a biginteger tambien

+++ database/migrations/2024_12_14_185730_create_assignments_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('assignments', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('id_teacher');
-            $table->unsignedInteger('id_module');
+            $table->unsignedBigInteger('id_teacher');
+            $table->unsignedBigInteger('id_module');
             $table->timestamps();
 
             $table->foreign('id_teacher')->references('id')->on('users')->onUpdate('cascade')->onDelete('cascade');
